split unknown_005740 into unknown_008C40 & unknown_00B8E0
